+++ DofusBot/Packet/Messages/Security/CheckIntegrityMessage.cs
@@ -14,12 +14,12 @@ namespace DofusBot.Packet.Messages.Security
         public override uint MessageID { get { return ProtocolId; } }
 
 
-        public List<byte> Data { get; set; }
+        public List<int> Data { get; set; }
 
 
         public CheckIntegrityMessage() { }
 
-        public CheckIntegrityMessage(List<byte> data)
+        public CheckIntegrityMessage(List<int> data)
         {
             Data = data;
         }
@@ -29,14 +29,14 @@ namespace DofusBot.Packet.Messages.Security
             writer.WriteVarInt(Data.Count);
             for (int i = 0; i < Data.Count; i++)
             {
-                writer.WriteByte(Data[i]);
+                writer.WriteByte((byte)Data[i]);
             }
         }
 
         public override void Deserialize(IDataReader reader)
         {
             int length = reader.ReadVarInt();
-            Data = new List<byte>();
+            Data = new List<int>();
             for (int i = 0; i < length; i++)
             {
                 Data.Add(reader.ReadByte());
